// npm/src/types/enums.ts
export enum TypeOfConstraint {
	AFTER_ALPHABETICALLY = 'AFTER_ALPHABETICALLY',
	AFTER_DATE = 'AFTER_DATE',
	AFTER_DATE_TIME = 'AFTER_DATE_TIME',
	AND = 'AND',
	AT_LEAST_N = 'AT_LEAST_N',
	AT_LEAST_N_CHARACTERS = 'AT_LEAST_N_CHARACTERS',
	AT_MOST_N = 'AT_MOST_N',
	AT_MOST_N_CHARACTERS = 'AT_MOST_N_CHARACTERS',
	BEFORE_ALPHABETICALLY = 'BEFORE_ALPHABETICALLY',
	BEFORE_DATE = 'BEFORE_DATE',
	BEFORE_DATE_TIME = 'BEFORE_DATE_TIME',
	CONFIRMED = 'CONFIRMED',
	DISJOINT_SET = 'DISJOINT_SET',
	EQUAL_TO_N = 'EQUAL_TO_N',
	EXACTLY_N_CHARACTERS = 'EXACTLY_N_CHARACTERS',
	FEWER_THAN_N_CHARACTERS = 'FEWER_THAN_N_CHARACTERS',
	IS_ARRAY = 'IS_ARRAY',
	IS_BOOLEAN = 'IS_BOOLEAN',
	IS_DATE = 'IS_DATE',
	IS_DATE_TIME = 'IS_DATE_TIME',
	IS_DURATION = 'IS_DURATION',
	IS_FRACTION = 'IS_FRACTION',
	IS_INSTANT = 'IS_INSTANT',
	IS_INTEGER = 'IS_INTEGER',
	IS_LIST = 'IS_LIST',
	IS_MAP = 'IS_MAP',
	IS_MEMBER = 'IS_MEMBER',
	IS_MONTH_DAY = 'IS_MONTH_DAY',
	IS_PRECISION = 'IS_PRECISION',
	IS_REAL = 'IS_REAL',
	IS_SET = 'IS_SET',
	IS_STRING = 'IS_STRING',
	IS_TIME_ZONE = 'IS_TIME_ZONE',
	IS_YEAR_MONTH = 'IS_YEAR_MONTH',
	IS_ZONED_DATE_TIME = 'IS_ZONED_DATE_TIME',
	LESS_THAN_N = 'LESS_THAN_N',
	MATCHING = 'MATCHING',
	MEMBER = 'MEMBER',
	MORE_THAN_N = 'MORE_THAN_N',
	MORE_THAN_N_CHARACTERS = 'MORE_THAN_N_CHARACTERS',
	NOT_AFTER_ALPHABETICALLY = 'NOT_AFTER_ALPHABETICALLY',
	NOT_BEFORE_ALPHABETICALLY = 'NOT_BEFORE_ALPHABETICALLY',
	NOT_EQUAL_TO_N = 'NOT_EQUAL_TO_N',
	ON_OR_AFTER_DATE = 'ON_OR_AFTER_DATE',
	ON_OR_BEFORE_DATE = 'ON_OR_BEFORE_DATE',
	OR = 'OR',
	ORDERED_LIST = 'ORDERED_LIST',
	OVERLAPPING_SET = 'OVERLAPPING_SET',
	REVERSED_LIST = 'REVERSED_LIST',
	SUBSET = 'SUBSET',
	SUPERSET = 'SUPERSET',
	XOR = 'XOR',
}